requirements.txt
asgiref==3.7.2
astroid==2.15.6
dill==0.3.7
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
Faker==19.3.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
platformdirs==3.10.0
pylint==2.17.5
python-dateutil==2.8.2
pytz==2020.1
six==1.15.0
sqlparse==0.4.4
toml==0.10.1
tomli==2.0.1
tomlkit==0.12.1
typed-ast==1.5.5
typing_extensions==4.7.1
validate-docbr==1.10.0
wrapt==1.12.1
